celery_worker/celery_worker.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()

REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")

celery_app = Celery("tasks", broker=REDIS_URL, backend=REDIS_URL)
